requirements.txt
requests==2.31.0
geopy==2.4.1
Pillow==10.2.0
customtkinter==5.2.2 